feat: add track and workshop registration views
Track and registration views were added as a first step towards implemementing the call for tracks application
together with the call for papers application. The menu for the call for tracks home is disabled but can be accessed
directly with the URL /cft/home. The persistence and form validation is not implemented