feat: status change to declined should not change the Approved or Refused collections in Redis
A proposal in the declined state can be in the Approved collections if it was accepted or approved and then declined or
in no collection if it was declined from the backup state. So if a proposal is changed to the declined state there
should be no change in the collections in the database. Previously it was added to the Refused collections.