fix: track scheduling contains talks that are not approved
The track scheduling contained talks that were not approved because the trackleader could cancel the approval after the talk was scheduled
Now when a talk is scheduled the button to cancel approval of the talk is disabled. If the trackleader really wishes to cancel the approval he needs to remove the talk from the schedule first.