[refactoring] change the application to read all configuration from the database
All the configuration specific to one conference have been removed from the ConferenceDescriptor and resource bundles, so they can be loaded from each conference from the database.
The only configuration still left hard-coded is the identifier for the conference that has to be loaded from the database. This is still configured as the event code in the ConferenceDescriptor class